requirements.txt
fastapi
uvicorn
nemo_toolkit[asr]
python-multipart
pydub
celery[redis]
redis
langchain_ollama
streamlit
requests
slowapi